random() and srandom() are POSIX; rand() and srand() are standard C.
BTW, you should never use the modulo operation to fit random values to
the desired range, as it skews the results towards the bottom of the
range.  For instance, if rand() returns a value between 0 and 255 and
you need a value between 0 and 9 (rand() % 10), 0 is 4% more likely to
appear than 9.